Renamed mockstore file
